+++ src/components/verification/ApprovalsList.js
@@ -1,42 +1,19 @@
-const ApprovalsList = ({ approvals }) => {
-  const approvalsMessages = {
-    undecided: (
-      <p className='font-bold text-[12px]'>
-        인증샷이 업로드되면 확인 후<br />
-        승인 또는 거절하실 수 있습니다
-      </p>
-    ),
-    approved: (
-      <p className='mb-1 font-semibold underline decoration-4 underline-offset-4 decoration-emerald-700'>
-        확인했어요!
-      </p>
-    ),
-    rejected: (
-      <p className='mb-1 font-semibold underline decoration-4 underline-offset-4 decoration-red-700'>
-        부족한 것 같아요..
-      </p>
-    ),
-  };
+import getUserIdFromToken from '../../utils/getUserIdFromToken';
+import ApprovalItem from './ApprovalItem';
 
-  const approvalsMessage = approvalsMessages['approved'];
+const ApprovalsList = ({ status, approvals }) => {
+  // const loginId = getUserIdFromToken();
+  const loginId = '6503cd53c7ffdf4851bb27d9';
+  const myApproval = approvals.filter((item) => item.userId._id === loginId)[0];
 
-  const profileImageUrl =
-    'https://lh3.googleusercontent.com/a/ACg8ocLX6SUyJF4Rg33r42HgO34BBwPJ5INmL_XWySlxDUAEyGkw=s96-c';
+  const updatedApprovals = approvals.filter((item) => item !== myApproval);
 
   return (
-    <div className='w-[240px] bg-dark-blue-bg rounded-xl'>
-      <div className='flex justify-center my-3'>
-        <div className='relative mr-[-25px] w-[40px] h-[40px] bg-light-gray-bg rounded-full flex'>
-          <img
-            src={profileImageUrl}
-            alt=''
-            className='object-cover rounded-full'
-          />
-        </div>
-        <div className='w-[210px] h-[43px] my-2 bg-dark-gray-bg rounded-full text-center flex items-center justify-center'>
-          {approvalsMessage}
-        </div>
-      </div>
+    <div className='w-[240px] bg-dark-blue-bg rounded-xl overflow-y-auto max-h-full transparent-scrollbar'>
+      {<ApprovalItem status={status} watcher={myApproval} />}
+      {updatedApprovals.map((watcher) => {
+        return <ApprovalItem status={status} watcher={watcher} />;
+      })}
     </div>
   );
 };
